+++ hpx/async_launch_policy_dispatch.hpp
@@ -8,15 +8,17 @@
 #define HPX_ASYNC_LAUNCH_POLICY_DISPATCH_NOV_26_2017_1243PM
 
 #include <hpx/config.hpp>
+#include <hpx/coroutines/thread_enums.hpp>
+#include <hpx/functional/deferred_call.hpp>
+#include <hpx/functional/invoke.hpp>
+#include <hpx/functional/traits/is_action.hpp>
 #include <hpx/lcos/async_fwd.hpp>
 #include <hpx/lcos/future.hpp>
 #include <hpx/lcos/local/futures_factory.hpp>
 #include <hpx/runtime/launch_policy.hpp>
 #include <hpx/runtime/threads/thread_data_fwd.hpp>
 #include <hpx/runtime/threads/thread_helpers.hpp>
-#include <hpx/functional/traits/is_action.hpp>
-#include <hpx/functional/deferred_call.hpp>
-#include <hpx/functional/invoke.hpp>
+#include <hpx/runtime/threads/thread_pool_base.hpp>
 
 #include <exception>
 #include <type_traits>
@@ -25,10 +27,10 @@
 namespace hpx { namespace detail
 {
     ///////////////////////////////////////////////////////////////////////////
-    template <typename F, typename ...Ts>
-    HPX_FORCEINLINE
-    lcos::future<typename util::detail::invoke_deferred_result<F, Ts...>::type>
-    call_sync(std::false_type, F f, Ts... vs) // decay-copy
+    template <typename F, typename... Ts>
+    HPX_FORCEINLINE lcos::future<
+        typename util::detail::invoke_deferred_result<F, Ts...>::type>
+    call_sync(std::false_type, F f, Ts... vs)    // decay-copy
     {
         using R = typename util::detail::invoke_deferred_result<F, Ts...>::type;
         try
@@ -38,15 +40,14 @@ namespace hpx { namespace detail
         }
         catch (...)
         {
-            return lcos::make_exceptional_future<R>(
-                std::current_exception());
+            return lcos::make_exceptional_future<R>(std::current_exception());
         }
     }
 
-    template <typename F, typename ...Ts>
-    HPX_FORCEINLINE
-    lcos::future<typename util::detail::invoke_deferred_result<F, Ts...>::type>
-    call_sync(std::true_type, F f, Ts... vs) // decay-copy
+    template <typename F, typename... Ts>
+    HPX_FORCEINLINE lcos::future<
+        typename util::detail::invoke_deferred_result<F, Ts...>::type>
+    call_sync(std::true_type, F f, Ts... vs)    // decay-copy
     {
         try
         {
@@ -63,35 +64,33 @@ namespace hpx { namespace detail
     ///////////////////////////////////////////////////////////////////////////
     template <typename Action>
     struct async_launch_policy_dispatch<Action,
-        typename std::enable_if<
-            !traits::is_action<Action>::value
-        >::type>
+        typename std::enable_if<!traits::is_action<Action>::value>::type>
     {
-        template <typename F, typename ...Ts>
-        HPX_FORCEINLINE static
-        typename std::enable_if<
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
             traits::detail::is_deferred_invocable<F, Ts...>::value,
-            hpx::future<
-                typename util::detail::invoke_deferred_result<F, Ts...>::type
-            >
-        >::type
-        call(launch policy, F && f, Ts&&... ts)
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(launch policy, threads::thread_pool_base* pool,
+            threads::thread_schedule_hint hint, F&& f, Ts&&... ts)
         {
-            typedef typename util::detail::invoke_deferred_result<F, Ts...>::type
-                result_type;
+            typedef
+                typename util::detail::invoke_deferred_result<F, Ts...>::type
+                    result_type;
 
             if (policy == launch::sync)
             {
                 using is_void = typename std::is_void<result_type>::type;
-                return detail::call_sync(is_void{},
-                    std::forward<F>(f), std::forward<Ts>(ts)...);
+                return detail::call_sync(
+                    is_void{}, std::forward<F>(f), std::forward<Ts>(ts)...);
             }
 
-            lcos::local::futures_factory<result_type()> p(
-                util::deferred_call(std::forward<F>(f), std::forward<Ts>(ts)...));
+            lcos::local::futures_factory<result_type()> p(util::deferred_call(
+                std::forward<F>(f), std::forward<Ts>(ts)...));
             if (hpx::detail::has_async_policy(policy))
             {
-                threads::thread_id_type tid = p.apply(policy, policy.priority());
+                threads::thread_id_type tid = p.apply(pool, policy,
+                    policy.priority(), threads::thread_stacksize_default, hint);
                 if (tid && policy == launch::fork)
                 {
                     // make sure this thread is executed last
@@ -103,63 +102,90 @@ namespace hpx { namespace detail
             return p.get_future();
         }
 
-        template <typename F, typename ...Ts>
-        HPX_FORCEINLINE static
-        typename std::enable_if<
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
             traits::detail::is_deferred_invocable<F, Ts...>::value,
-            hpx::future<
-                typename util::detail::invoke_deferred_result<F, Ts...>::type
-            >
-        >::type
-        call(hpx::detail::sync_policy, F && f, Ts&&... ts)
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(launch policy, F&& f, Ts&&... ts)
         {
-            typedef typename util::detail::invoke_deferred_result<F, Ts...>::type
-                result_type;
-
-            using is_void = typename std::is_void<result_type>::type;
-            return detail::call_sync(is_void{},
-                std::forward<F>(f), std::forward<Ts>(ts)...);
+            return call(policy, threads::detail::get_self_or_default_pool(),
+                threads::thread_schedule_hint{}, std::forward<F>(f),
+                std::forward<Ts>(ts)...);
         }
 
-        template <typename F, typename ...Ts>
-        HPX_FORCEINLINE static
-        typename std::enable_if<
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
             traits::detail::is_deferred_invocable<F, Ts...>::value,
-            hpx::future<
-                typename util::detail::invoke_deferred_result<F, Ts...>::type
-            >
-        >::type
-        call(hpx::detail::async_policy policy, F && f, Ts&&... ts)
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(hpx::detail::sync_policy, F&& f, Ts&&... ts)
         {
-            typedef typename util::detail::invoke_deferred_result<F, Ts...>::type
-                result_type;
+            typedef
+                typename util::detail::invoke_deferred_result<F, Ts...>::type
+                    result_type;
 
-            lcos::local::futures_factory<result_type()> p(
-                util::deferred_call(std::forward<F>(f), std::forward<Ts>(ts)...));
+            using is_void = typename std::is_void<result_type>::type;
+            return detail::call_sync(
+                is_void{}, std::forward<F>(f), std::forward<Ts>(ts)...);
+        }
 
-            p.apply(policy, policy.priority());
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
+            traits::detail::is_deferred_invocable<F, Ts...>::value,
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(hpx::detail::async_policy policy, threads::thread_pool_base* pool,
+            threads::thread_schedule_hint hint, F&& f, Ts&&... ts)
+        {
+            HPX_ASSERT(pool);
+            typedef
+                typename util::detail::invoke_deferred_result<F, Ts...>::type
+                    result_type;
+
+            lcos::local::futures_factory<result_type()> p(util::deferred_call(
+                std::forward<F>(f), std::forward<Ts>(ts)...));
+
+            p.apply(pool, policy, policy.priority(),
+                threads::thread_stacksize_default, hint);
             return p.get_future();
         }
 
-        template <typename F, typename ...Ts>
-        HPX_FORCEINLINE static
-        typename std::enable_if<
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
             traits::detail::is_deferred_invocable<F, Ts...>::value,
-            hpx::future<
-                typename util::detail::invoke_deferred_result<F, Ts...>::type
-            >
-        >::type
-        call(hpx::detail::fork_policy policy, F && f, Ts&&... ts)
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(hpx::detail::async_policy policy, F&& f, Ts&&... ts)
         {
-            typedef typename util::detail::invoke_deferred_result<F, Ts...>::type
-                result_type;
+            return call(policy, threads::detail::get_self_or_default_pool(),
+                threads::thread_schedule_hint{}, std::forward<F>(f),
+                std::forward<Ts>(ts)...);
+        }
 
-            lcos::local::futures_factory<result_type()> p(
-                util::deferred_call(std::forward<F>(f), std::forward<Ts>(ts)...));
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
+            traits::detail::is_deferred_invocable<F, Ts...>::value,
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(hpx::detail::fork_policy policy, threads::thread_pool_base* pool,
+            threads::thread_schedule_hint hint, F&& f, Ts&&... ts)
+        {
+            HPX_ASSERT(pool);
+            typedef
+                typename util::detail::invoke_deferred_result<F, Ts...>::type
+                    result_type;
+
+            lcos::local::futures_factory<result_type()> p(util::deferred_call(
+                std::forward<F>(f), std::forward<Ts>(ts)...));
 
             // make sure this thread is executed last
-            threads::thread_id_type tid = p.apply(policy, policy.priority());
-            if (tid)
+            threads::thread_id_type tid = p.apply(pool, policy,
+                policy.priority(), threads::thread_stacksize_default, hint);
+            threads::thread_id_type tid_self = threads::get_self_id();
+            if (tid && tid_self &&
+                get_thread_id_data(tid)->get_scheduler_base() ==
+                    get_thread_id_data(tid_self)->get_scheduler_base())
             {
                 // yield_to
                 hpx::this_thread::suspend(threads::pending, tid,
@@ -168,25 +194,35 @@ namespace hpx { namespace detail
             return p.get_future();
         }
 
-        template <typename F, typename ...Ts>
-        HPX_FORCEINLINE static
-        typename std::enable_if<
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
             traits::detail::is_deferred_invocable<F, Ts...>::value,
-            hpx::future<
-                typename util::detail::invoke_deferred_result<F, Ts...>::type
-            >
-        >::type
-        call(hpx::detail::deferred_policy, F && f, Ts &&... ts)
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(hpx::detail::fork_policy policy, F&& f, Ts&&... ts)
         {
-            typedef typename util::detail::invoke_deferred_result<F, Ts...>::type
-                result_type;
+            return call(policy, threads::detail::get_self_or_default_pool(),
+                threads::thread_schedule_hint{}, std::forward<F>(f),
+                std::forward<Ts>(ts)...);
+        }
 
-            lcos::local::futures_factory<result_type()> p(
-                util::deferred_call(std::forward<F>(f), std::forward<Ts>(ts)...));
+        template <typename F, typename... Ts>
+        HPX_FORCEINLINE static typename std::enable_if<
+            traits::detail::is_deferred_invocable<F, Ts...>::value,
+            hpx::future<typename util::detail::invoke_deferred_result<F,
+                Ts...>::type>>::type
+        call(hpx::detail::deferred_policy, F&& f, Ts&&... ts)
+        {
+            typedef
+                typename util::detail::invoke_deferred_result<F, Ts...>::type
+                    result_type;
+
+            lcos::local::futures_factory<result_type()> p(util::deferred_call(
+                std::forward<F>(f), std::forward<Ts>(ts)...));
 
             return p.get_future();
         }
     };
-}}
+}}    // namespace hpx::detail
 
 #endif
