+++ hpx/lcos/local/packaged_continuation.hpp
@@ -382,7 +382,7 @@ namespace hpx { namespace lcos { namespace detail
 
             parallel::execution::detail::post_policy_dispatch<
                     hpx::launch::async_policy
-                >::call(desc, hpx::launch::async,
+                >::call(hpx::launch::async, desc,
                     [HPX_CAPTURE_MOVE(this_),
                         HPX_CAPTURE_MOVE(f)
                     ]() mutable -> void {
@@ -428,7 +428,7 @@ namespace hpx { namespace lcos { namespace detail
 
             parallel::execution::detail::post_policy_dispatch<
                     hpx::launch::async_policy
-                >::call(desc, hpx::launch::async,
+                >::call(hpx::launch::async, desc,
                     [HPX_CAPTURE_MOVE(this_),
                         HPX_CAPTURE_MOVE(f)
                     ]() mutable -> void {
