+++ hpx/runtime/threads/detail/scheduling_loop.hpp
@@ -12,6 +12,7 @@
 #include <hpx/runtime/get_thread_name.hpp>
 #include <hpx/runtime/threads/policies/scheduler_base.hpp>
 #include <hpx/runtime/threads/thread_data.hpp>
+#include <hpx/runtime/threads/thread_helpers.hpp>
 #if defined(HPX_HAVE_BACKGROUND_THREAD_COUNTERS) && defined(HPX_HAVE_THREAD_IDLE_RATES)
 # include <hpx/runtime/threads/scoped_background_timer.hpp>
 #endif
